fix(broker): keep fractional quantity for BUY swaps
quantity = int(order.quantity) truncated 0.05 → 0, so the BUY sizing
formula produced --readable-amount 0.000000 and onchainos CLI aborted
('too small for this token (6 decimals); results in zero minimal units').
Use float() so explicit fractional quantity (e.g. 0.05 CRCLX) is passed
through; the pipeline-level _calculate_quantity default (int + min 1)
is unchanged. tests: 265 passed.

+++ src/nanobot_quant/brokers/onchainos_broker.py
@@ -87,7 +87,10 @@ class OnchainOSBroker(Broker):
         """
         symbol = order.asset.symbol
         side = order.side.lower()
-        quantity = int(order.quantity)
+        # Keep decimals: explicit quantity (e.g. 0.05 CRCLX) must not be
+        # truncated to 0 — int() turned 0.05 into 0 → swap aborted with
+        # "--readable-amount 0.000000 is too small for this token".
+        quantity = float(order.quantity)
 
         if side not in ("buy", "sell"):
             order.set_error(f"Unsupported side: {side} (only buy/sell)")
